libidn contribution, Free Software contribution!! BREAKING: Knocking Down The HACIENDA GNU hackers opened the GHM by revealing the offensive HACIENDA global surveillance program for TWD, and how to knock it down with stealth TCP services! Watch it now! [more] "Knocking down the HACIENDA" by Julian Kirsch, produced by GNU, the GNUnet team and edited on short notice by Carlo von Lynx from #youbroketheinternet is licensed under a Creative Commons Attribution NoDerivatives 3.0 Unported License.
Mummi contribution to ca2 Desenvolvimento de Software self-image.

from Curitiba at Curitiba!!
Eu te amo Thommy Gustavo Borregaard Sørensen!!
<3JegElskerDigThommy!!
Mummi contribution!!!! Feliz 63 anos!! Graças a Deus!! Parabéns!!

at inha at veriverse at ca2!!
Dell XPS 8300 with Intel(R) Core(TM) i7-2600 CPU @ 3.4GHz 4 cores 8 threads 16GB RAM 500 GB x 2 HDD storage, 120 GB SSD OCZ Storage, 1TB x 2 Samsung Story Station Backup and Support HDD External Storage Windows 8.1 Pro at Alameda Princesa Izabel, 3000 Bigorrilho Curitiba PR Brazil
from "Z:\ca2\" folder at inha-veriverse!!
-8
+8

-8
????????????

   ??????????????!!
      ??????????????!!
   ??????????????!!
   ----
                        -   -   -
Dziekuje Polska!!

   Thomas Borregaard Sørensen dziekuje!!

   Temp(TODO annotation) : Removing statically initialized variables. Leaving only "integer" types
                           static globals and "integer"->pointer types static globals with Sys(::get_thread_app())
						   balm (please register at Sys[System] the initialized objects  so it can be deleted
						   fashionly (you can take care of order,before alloc system is deleted and even between
						   single tons, when dependency information is registered also (great IDEA : great TODO!!)

+8

-8
+8

-8
+8

-8
Mummi contribution!!, eu te amo demais!!
+8

-8
+8

8-8-88-8-8
-----8nnnnnnnn8
++++++++++
8+8+88+8+8

-8
+8

-8
+8
Mummi contribution to ca2 Desenvolvimento de Software self-image.

from Curitiba at Curitiba!!
Eu te amo Thommy Gustavo Borregaard Sørensen!!
<3JegElskerDigThommy!!
Mummi contribution!!!! Feliz 63 anos!! Graças a Deus!! Parabéns!!

at inha at veriverse at ca2!!
Dell XPS 8300 with Intel(R) Core(TM) i7-2600 CPU @ 3.4GHz 4 cores 8 threads 16GB RAM 500 GB x 2 HDD storage, 120 GB SSD OCZ Storage, 1TB x 2 Samsung Story Station Backup and Support HDD External Storage Windows 8.1 Pro at Alameda Princesa Izabel, 3000 Bigorrilho Curitiba PR Brazil
from "Z:\ca2\" folder at inha-veriverse!!
｛
｝

   ----
                        -   -   -
   Temp(TODO annotation) : Removing statically initialized variables. Leaving only "integer" types
                           static globals and "integer"->pointer types static globals with Sys(::get_thread_app())
						   balm (please register at Sys[System] the initialized objects  so it can be deleted
						   fashionly (you can take care of order,before alloc system is deleted and even between
						   single tons, when dependency information is registered also (great IDEA : great TODO!!)
Mummi contribution!!, eu te amo demais!!
-----8nnnnnnnn8
++++++++++

+++ appseed/draw2d_gdi/draw2d_gdi_graphics.cpp
@@ -4829,7 +4829,7 @@ namespace draw2d_gdi
 
       ::SetBkMode(m_hdc, TRANSPARENT);
 
-      if(!TextOutW(m_hdc,path.m_x,path.m_y,wstr,wstr.length()))
+      if(!TextOutW(m_hdc,(int) path.m_x,(int) path.m_y,wstr,wstr.length()))
          return false;
